Scope config support, resource type extension
We prefix file downloads with the resource type that is being downloaded
and show the type in the filename being downloaded.  Since some FHIR
servers (such as OpenEMR) uses the content-disposition and file-name
property to specify the name of a download and the fhir downloader
does not respect that choice we had to differentiate the files so that everything
doesn't get overwritten.

Also some FHIR auth servers (such as OpenEMR) do not allow global
permissions ie (system/*.*) and permissions have to be specified
manually.  We modified the tool so that it can specify scopes on both
the CLI as well as in the config file.